Move ChartBuilderMenuItem.tsx to /internal/components/chart
